+++ src/app/mod.rs
@@ -2,6 +2,7 @@ mod account;
 mod accounts;
 mod chart;
 mod crypto;
+mod crypto2;
 mod import_boa;
 mod message;
 mod metal;